% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_term_long.R
\name{pivot_term_long}
\alias{pivot_term_long}
\title{Pivot a term (sentence) into a data.frame (long)}
\usage{
pivot_term_long(term)
}
\arguments{
\item{term}{string, term, sentence, etc.}
}
\value{
data.frame with \verb{Unique Items} and \code{Term}
}
\description{
Pivot a term (sentence) into a data.frame (long)
}
\examples{
pivot_term_long("A large size in stockings is hard to sell.")
}
